💄Style: Articleに関する共通コンポーネントを修正
